pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
description = "none"
version = "0.0.0"
[tool.poetry.dependencies]
python = "^3.8"
python-chess = "*"
